Restructuring Spring Boot application
It is recommend that you locate your main application class in a root package above other classes.
