+++ AxonPOC_Demo/src/test/java/com/axonpoc/AxonPOC_DemoApplicationTests.java
@@ -1,0 +1,36 @@
+
+package com.axonpoc;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+/**
+ * Copyright (C) 2006 by Citicorp Development Center, Inc. All rights
+ * reserved. Citicorp Development Center, Inc. claims copyright in
+ * this computer program as an unpublished work, one or more versions
+ * of which were first used to provide services to customers on the
+ * dates indicated in the foregoing notice. Claim of copyright does
+ * not imply waiver of other rights
+ * <p>
+ * NOTICE OF PROPRIETARY RIGHTS
+ * <p>
+ * This program is a confidential trade secret and the property of
+ * Citicorp Development Center, Inc. Use, examination, reproduction,
+ * disassembly, decompiling, transfer and/or disclosure to others
+ * of all or any part of this software program are strictly prohibited
+ * except by express written agreement with Citicorp Development
+ * Center, Inc.
+ *
+ * @author Citi
+ */
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class AxonPOC_DemoApplicationTests {
+
+    @Test
+    public void contextLoads() {
+    }
+}
